+++ db/migrate/20250504144221_create_llm_requests.rb
@@ -1,0 +1,15 @@
+class CreateLlmRequests < ActiveRecord::Migration[8.0]
+  def change
+    create_table :llm_requests do |t|
+      t.references :trackable, polymorphic: true, null: false
+      t.references :user, null: false, foreign_key: true
+      t.integer :llm
+      t.integer :request_type
+      t.integer :token_count
+      t.integer :micro_usd
+      t.text :prompt
+
+      t.timestamps
+    end
+  end
+end